date showing fixed for new tree
